pyproject.toml
[project]
name = "twitscraper"
version = "0.1.0"
description = "A simple Twitter scraper"
authors = ["NotCoderGuy <me@notcoderguy.com>"]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.0",
    "twikit>=2.3.3",
    "pyotp>=2.9.0",
    "selenium>=4.0.0",
    "webdriver-manager>=3.0.0",
]
